moved celery test case base class
