+++ djcelery_model/tests/test_base.py
@@ -1,0 +1,10 @@
+from celery.contrib.testing.tasks import ping
+
+from .base import CeleryTestCase
+
+
+class TestCeleryTestCase(CeleryTestCase):
+    def test_worker(self):
+        result = ping.delay()
+        pong = result.get(timeout=10)
+        self.assertEqual(pong, 'pong')
